Improve SRP by decomposing Utils.java and CSVImporter.java
